[5G-SA][lte][agw][SMF]UPF Node message Handling
